Fix doctest

+++ cupy/indexing/generate.py
@@ -114,7 +114,7 @@ Examples
 >>> a = cupy.array([1, 2, 3], dtype=np.int32)
 >>> b = cupy.array([4, 5, 6], dtype=np.int32)
 >>> cupy.c_[a, 0, 0, b]
-array([[1, 2, 3, 0, 0, 4, 5, 6]])
+array([[1, 2, 3, 0, 0, 4, 5, 6]], dtype=int32)
 
 """
 
@@ -152,7 +152,7 @@ Examples
 >>> a = cupy.array([1, 2, 3], dtype=np.int32)
 >>> b = cupy.array([4, 5, 6], dtype=np.int32)
 >>> cupy.r_[a, 0, 0, b]
-array([1, 2, 3, 0, 0, 4, 5, 6])
+array([1, 2, 3, 0, 0, 4, 5, 6], dtype=int32)
 
 """
